+++ can/tests/test_dispatch.cpp
@@ -1,11 +1,13 @@
 #include "can/core/dispatch.hpp"
 #include "can/core/messages.hpp"
+#include "can/core/arbitration_id.hpp"
 #include "can/tests/mock_message_buffer.hpp"
 #include "catch2/catch.hpp"
 #include "common/core/bit_utils.hpp"
 
 using namespace can_dispatch;
 using namespace can_messages;
+using namespace can_arbitration_id;
 using namespace mock_message_buffer;
 
 SCENARIO("Dispatcher") {
