+++ include/can/core/ids.hpp
@@ -5,24 +5,6 @@
 namespace can_ids {
 
 /**
- * The components of a 29-bit arbitration id as a bitfield.
- */
-struct ArbitrationIdParts {
-    unsigned function_code : 7;
-    unsigned node_id : 8;
-    unsigned message_id : 14;
-    unsigned int padding : 3;
-};
-
-/**
- * A union of arbitration id in parts or as an integer.
- */
-union ArbitrationId {
-    ArbitrationIdParts parts;
-    uint32_t id;
-};
-
-/**
  * Function code definitions.
  */
 enum class FunctionCode : uint8_t {
